fix(books): move ?immersive=1 handling out of provider for static export
CI's static-export build failed on the v3 polish branch with:

  useSearchParams() should be wrapped in a suspense boundary at
  page "/books/[slug]/[...chapter]"

The auto-enter logic that reads `?immersive=1` from the URL lived
inside ImmersiveReadingProvider via useSearchParams. Next.js
static export requires any useSearchParams caller to sit under a
Suspense boundary so it can defer to client rendering. Wrapping
the whole provider would also drag {children} (the chapter page)
into the suspended subtree, losing prerender.

Fix: extract the flag-reading logic into a new
ImmersiveReadingFlagHandler client component that renders null,
wrap only that component in <Suspense fallback={null}> inside the
books/[slug] layout, and leave the provider + chapter siblings
outside the boundary. The handler runs after hydration, calls
provider.enter(), and replaces the URL with the flag stripped —
same behaviour as before, just relocated.

Verified locally with `bun run build:dev` — 999/999 pages
prerender cleanly.
